Random numbers generated needed fix
Random.nextInt(int n) generates a random number in the range 0(inclusive) to n(exclusive).
We want to generate index for arrays flavors and UPCs randomly. So we need indexes from
0 to flavors.length-1 (both inclusive). So for upper bound we will have to add 1 to the upper bound
and the pass this to nextInt() method. So, correct call should be nextInt(flavors.length), then only it
will generate number from 0 to flavors.length-1 (both inclusive).