What I did:
I edited the code and I switched "New Coke" with no value to Dr.Pepper
that has value and now I get output for Dr.Pepper to be added to the
stack. I also added a parameter to the class as well to ensure that
Dr.Pepper is added to the stack.